fix(client): add language switch button outside login (fixes "Add language switch before login page" #46)
